+++ src/renderer/DiagnosticsDialog.tsx
@@ -6,7 +6,7 @@ import {
   DialogTrigger,
 } from './components/Dialog/Dialog';
 import { Button } from './components/Button/Button';
-import { ReactNode, useState } from 'react';
+import { Dispatch, ReactNode, SetStateAction, useState } from 'react';
 import { AppState } from 'main/types';
 import { ExternalLink, FileText, Package } from 'lucide-react';
 import Spinner from './components/Spinner/Spinner';
@@ -16,6 +16,7 @@ import { getLocalePhrase } from 'localisation/translations';
 type IProps = {
   children: ReactNode;
   appState: AppState;
+  setPreviewEnabled: Dispatch<SetStateAction<boolean>>;
 };
 
 const ipc = window.electron.ipcRenderer;
@@ -25,7 +26,7 @@ const openLogPath = () => {
 };
 
 const DiagnosticsDialog = (props: IProps) => {
-  const { children, appState } = props;
+  const { children, appState, setPreviewEnabled } = props;
   const { language } = appState;
   const [open, setOpen] = useState(false);
   const [zipping, setZipping] = useState(false);
@@ -105,6 +106,7 @@ const DiagnosticsDialog = (props: IProps) => {
       open={open}
       onOpenChange={(open) => {
         setOpen(open);
+        setPreviewEnabled(!open);
         if (open) {
           setZipping(false);
           setBundlePath('');